requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
argilla==1.6.0
async-timeout==4.0.2
attrs==21.2.0
Automat==20.2.0
autopep8==2.0.2
awscli==1.22.34
Babel==2.8.0
backoff==2.2.1
bcrypt==3.2.0
beautifulsoup4==4.12.2
blinker==1.6.2
botocore==1.23.34
bs4==0.0.1
cachetools==5.3.0
certifi==2022.12.7
chardet==4.0.0
charset-normalizer==3.1.0
cliapp==1.20180812.1
click==8.1.3
cloud-init==23.1.2
colorama==0.4.4
command-not-found==0.3
commonmark==0.9.1
configobj==5.0.6
constantly==15.1.0
cryptography==3.4.8
dataclasses-json==0.5.7
dbus-python==1.2.18
Deprecated==1.2.13
distlib==0.3.4
distro==1.7.0
distro-info===1.1build1
dnspython==2.3.0
docutils==0.17.1
dropbox==11.36.0
ec2-hibinit-agent==1.0.0
et-xmlfile==1.1.0
filelock==3.6.0
Flask==2.3.2
Flask-Cors==3.0.10
frozenlist==1.3.3
google-api-core==2.11.0
google-api-python-client==2.83.0
google-auth==2.17.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httplib2==0.20.2
httpx==0.23.3
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.2.0
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
langchain==0.0.155
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
loguru==0.7.0
lxml==4.9.2
Markdown==3.3.6
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
monotonic==1.6
more-itertools==8.10.0
msg-parser==1.2.0
multidict==6.0.4
mypy-extensions==1.0.0
netifaces==0.11.0
nltk==3.8.1
numexpr==2.8.4
numpy==1.23.5
oauthlib==3.2.0
olefile==0.46
openai==0.27.6
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
packaging==23.1
pandas==1.5.3
pexpect==4.8.0
Pillow==9.0.1
pinecone-client==2.2.1
platformdirs==2.5.1
ply==3.11
protobuf==4.22.1
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycodestyle==2.10.0
pydantic==1.10.7
Pygments==2.11.2
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pyOpenSSL==21.0.0
pypandoc==1.11
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
python-docx==0.8.11
python-dotenv==1.0.0
python-magic==0.4.24
python-pptx==0.6.21
pytz==2022.1
PyYAML==5.4.1
regex==2023.3.23
requests==2.25.1
rfc3986==1.5.0
rich==13.0.1
roman==3.3
rsa==4.9
s3transfer==0.5.0
SecretStorage==3.3.1
service-identity==18.1.0
six==1.16.0
sniffio==1.3.0
sos==4.4
soupsieve==2.4.1
SQLAlchemy==2.0.12
ssh-import-id==5.11
stone==3.3.1
systemd-python==234
tenacity==8.2.2
tomli==2.0.1
tqdm==4.65.0
ttystatus==0.38
Twisted==22.1.0
typing-inspect==0.8.0
typing_extensions==4.5.0
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
unstructured==0.6.2
uritemplate==4.1.1
urllib3==1.26.5
virtualenv==20.13.0+ds
wadllib==1.3.6
Werkzeug==2.3.4
wrapt==1.14.1
XlsxWriter==3.1.0
yarl==1.9.2
zipp==1.0.0
zope.interface==5.4.0
